+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useChainInfo';
+export * from './useAuth3Token';